backend/requirements/base.txt
Django==3.2.13
djangorestframework==3.13.1
django-model-utils==4.2.0